Add db index pair constraint for answers table
Two commits ago (3242f9d), I added model validations so there couldn't
be 2 answers with the same question_id and team_id, because it would
mean the same team answered the same question twice (or more) instead of
updating their existing answer. But, as explained in Rails'
documentation, there could still exist the possibility of a collision
through a race condition.

The solution offered there was to add unique constraints to the
database. So I added them, Now it doesn't matter if a race condition
occurs, the database will assure the (team_id, question_id) tuple is
still unique.