Simplify questions views
Improved some details in the questions controller, p.e. removing the
`@team`, which is redundant when there already exists `current_team`,
but most of the improvements/simplifications are in the views.

It should be noted that I added `Question.last.id` directly into the
`question#show` view. Since I'm planning on making a json builder, it'd
be useless to add a '@question_last_id' variable in `quesiton#show'.
Hope it's the right decision.